+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/base/BaseOperation.java
@@ -51,6 +51,7 @@ import io.fabric8.kubernetes.client.dsl.internal.DefaultOperationInfo;
 import io.fabric8.kubernetes.client.dsl.internal.WatchConnectionManager;
 import io.fabric8.kubernetes.client.dsl.internal.WatchHTTPManager;
 import io.fabric8.kubernetes.client.internal.readiness.Readiness;
+import io.fabric8.kubernetes.client.utils.HttpClientUtils;
 import io.fabric8.kubernetes.client.utils.URLUtils;
 import io.fabric8.kubernetes.client.utils.Utils;
 import io.fabric8.kubernetes.client.utils.WatcherToggle;
@@ -171,35 +172,7 @@ public class BaseOperation<T extends HasMetadata, L extends KubernetesResourceLi
   }
 
   protected URL fetchListUrl(URL url, ListOptions listOptions) throws MalformedURLException {
-    HttpUrl.Builder urlBuilder = HttpUrl.get(url.toString()).newBuilder();
-    if(listOptions.getLimit() != null) {
-      urlBuilder.addQueryParameter("limit", listOptions.getLimit().toString());
-    }
-    if(listOptions.getContinue() != null) {
-      urlBuilder.addQueryParameter("continue", listOptions.getContinue());
-    }
-
-    if (listOptions.getResourceVersion() != null) {
-      urlBuilder.addQueryParameter("resourceVersion", listOptions.getResourceVersion());
-    }
-
-    if (listOptions.getFieldSelector() != null) {
-      urlBuilder.addQueryParameter("fieldSelector", listOptions.getFieldSelector());
-    }
-
-    if (listOptions.getLabelSelector() != null) {
-      urlBuilder.addQueryParameter("labelSelector", listOptions.getLabelSelector());
-    }
-
-    if (listOptions.getTimeoutSeconds() != null) {
-      urlBuilder.addQueryParameter("timeoutSeconds", listOptions.getTimeoutSeconds().toString());
-    }
-
-    if (listOptions.getAllowWatchBookmarks() != null) {
-      urlBuilder.addQueryParameter("allowWatchBookmarks", listOptions.getAllowWatchBookmarks().toString());
-    }
-
-    return new URL(urlBuilder.toString());
+    return new URL(HttpClientUtils.appendListOptionParams(HttpUrl.get(url.toString()).newBuilder(), listOptions).toString());
   }
 
   private void addQueryStringParam(HttpUrl.Builder requestUrlBuilder, String name, String value) {
@@ -319,7 +292,7 @@ public class BaseOperation<T extends HasMetadata, L extends KubernetesResourceLi
 
   @Override
   public EditReplacePatchDeletable<T, T, D, Boolean> cascading(boolean cascading) {
-    return newInstance(context.withCascading(cascading));
+    return newInstance(context.withCascading(cascading).withPropagationPolicy(null));
   }
 
   @Override
@@ -779,18 +752,29 @@ public class BaseOperation<T extends HasMetadata, L extends KubernetesResourceLi
     return newInstance(context.withResourceVersion(resourceVersion));
   }
 
-  public Watch watch(final Watcher<T> watcher) throws KubernetesClientException {
-    return watch(resourceVersion, watcher);
+  public Watch watch(final Watcher<T> watcher) {
+    return watch(new ListOptionsBuilder()
+      .withResourceVersion(resourceVersion)
+      .build(), watcher);
   }
 
-  public Watch watch(String resourceVersion, final Watcher<T> watcher) throws KubernetesClientException {
+  @Override
+  public Watch watch(String resourceVersion, Watcher<T> watcher) {
+    return watch(new ListOptionsBuilder()
+      .withResourceVersion(resourceVersion)
+      .build(), watcher);
+  }
+
+  @Override
+  public Watch watch(ListOptions options, final Watcher<T> watcher) {
     WatcherToggle<T> watcherToggle = new WatcherToggle<>(watcher, true);
+    options.setWatch(Boolean.TRUE);
     WatchConnectionManager watch = null;
     try {
       watch = new WatchConnectionManager(
         client,
         this,
-        resourceVersion,
+        options,
         watcherToggle,
         config.getWatchReconnectInterval(),
         config.getWatchReconnectLimit(),
@@ -824,7 +808,7 @@ public class BaseOperation<T extends HasMetadata, L extends KubernetesResourceLi
         return new WatchHTTPManager(
           client,
           this,
-          resourceVersion,
+          options,
           watcher,
           config.getWatchReconnectInterval(),
           config.getWatchReconnectLimit(),
@@ -846,8 +830,9 @@ public class BaseOperation<T extends HasMetadata, L extends KubernetesResourceLi
     throw new KubernetesClientException("Cannot update read-only resources");
   }
 
+  @Override
   public boolean isResourceNamespaced() {
-    return true;
+    return Utils.isResourceNamespaced(getType());
   }
 
   protected T handleResponse(Request.Builder requestBuilder) throws ExecutionException, InterruptedException, KubernetesClientException, IOException {