+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/WatchHTTPManager.java
@@ -19,6 +19,7 @@ package io.fabric8.kubernetes.client.dsl.internal;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.KubernetesResource;
 import io.fabric8.kubernetes.api.model.KubernetesResourceList;
+import io.fabric8.kubernetes.api.model.ListOptions;
 import io.fabric8.kubernetes.api.model.Status;
 import io.fabric8.kubernetes.api.model.WatchEvent;
 import io.fabric8.kubernetes.client.KubernetesClientException;
@@ -27,6 +28,7 @@ import io.fabric8.kubernetes.client.Watcher;
 import io.fabric8.kubernetes.client.Watcher.Action;
 import io.fabric8.kubernetes.client.dsl.base.BaseOperation;
 import io.fabric8.kubernetes.client.dsl.base.OperationSupport;
+import io.fabric8.kubernetes.client.utils.HttpClientUtils;
 import io.fabric8.kubernetes.client.utils.Serialization;
 import io.fabric8.kubernetes.client.utils.Utils;
 import okhttp3.*;
@@ -53,6 +55,7 @@ public class WatchHTTPManager<T extends HasMetadata, L extends KubernetesResourc
   private final BaseOperation<T, L, ?, ?> baseOperation;
   private final Watcher<T> watcher;
   private final AtomicBoolean forceClosed = new AtomicBoolean();
+  private final ListOptions listOptions;
   private final AtomicReference<String> resourceVersion;
   private final int reconnectLimit;
   private final int reconnectInterval;
@@ -71,20 +74,21 @@ public class WatchHTTPManager<T extends HasMetadata, L extends KubernetesResourc
 
   public WatchHTTPManager(final OkHttpClient client,
                           final BaseOperation<T, L, ?, ?> baseOperation,
-                          final String version, final Watcher<T> watcher, final int reconnectInterval,
+                          final ListOptions listOptions, final Watcher<T> watcher, final int reconnectInterval,
                           final int reconnectLimit, long connectTimeout)
     throws MalformedURLException {
     // Default max 32x slowdown from base interval
-    this(client, baseOperation, version, watcher, reconnectInterval, reconnectLimit, connectTimeout, 5);
+    this(client, baseOperation, listOptions, watcher, reconnectInterval, reconnectLimit, connectTimeout, 5);
   }
 
   public WatchHTTPManager(final OkHttpClient client,
                           final BaseOperation<T, L, ?, ?> baseOperation,
-                          final String version, final Watcher<T> watcher, final int reconnectInterval,
+                          final ListOptions listOptions, final Watcher<T> watcher, final int reconnectInterval,
                           final int reconnectLimit, long connectTimeout, int maxIntervalExponent)
     throws MalformedURLException {
 
-    this.resourceVersion = new AtomicReference<>(version); // may be a reference to null
+    this.resourceVersion = new AtomicReference<>(listOptions.getResourceVersion());
+    this.listOptions = listOptions;
     this.baseOperation = baseOperation;
     this.watcher = watcher;
     this.reconnectInterval = reconnectInterval;
@@ -140,12 +144,8 @@ public class WatchHTTPManager<T extends HasMetadata, L extends KubernetesResourc
       httpUrlBuilder.addQueryParameter("fieldSelector", fieldQueryString);
     }
 
-    if (this.resourceVersion.get() != null) {
-      httpUrlBuilder.addQueryParameter("resourceVersion", this.resourceVersion.get());
-    }
-
-    httpUrlBuilder.addQueryParameter("watch", "true");
-
+    listOptions.setResourceVersion(resourceVersion.get());
+    HttpClientUtils.appendListOptionParams(httpUrlBuilder, listOptions);
     String origin = requestUrl.getProtocol() + "://" + requestUrl.getHost();
     if (requestUrl.getPort() != -1) {
         origin += ":" + requestUrl.getPort();