+++ kubernetes-model/kubernetes-model-common/src/main/java/io/fabric8/kubernetes/model/annotation/ApiVersion.java
@@ -16,6 +16,10 @@
 
 package io.fabric8.kubernetes.model.annotation;
 
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
 public @interface ApiVersion {
 
   String value();