+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/crd/StarList.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.kubernetes.model.annotation;
+package io.fabric8.kubernetes.client.mock.crd;
 
-public @interface Namespaced {
-  boolean value();
+import io.fabric8.kubernetes.client.CustomResourceList;
+
+public class StarList extends CustomResourceList<Star> {
 }